+++ packages/util-crypto/src/base64/decode.spec.ts
@@ -8,6 +8,14 @@ import { stringToU8a } from '@polkadot/util';
 import { base64Decode } from './index.js';
 
 describe('base64Decode', (): void => {
+  it('decodes an empty string)', (): void => {
+    expect(
+      base64Decode('')
+    ).toEqual(
+      stringToU8a('')
+    );
+  });
+
   it('decodes a mixed base64 utf8 string (1)', (): void => {
     expect(
       base64Decode('aGVsbG8gd29ybGQg0J/RgNC40LLQtdGC0YHRgtCy0YPRjiDQvNC4IOS9oOWlvQ==')